+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Auto Elemental` OP abilities (requires TabletopTweaks-Base).
 - Added UMM option to allow for multiple Mythic OP abilities and Mythic special powers.
 - Added UMM option to restrict Exceptional Feats to the Isekai Protagonist Class.
+- Added UMM option to restrict Mythic Overpowered Abilities to the Isekai Protagonist Class.
+- Added UMM option to restrict Mythic Special Powers to the Isekai Protagonist Class.
 - Updated mod support for ExpandedContent.
 ### Changed
 - Reworked UMM settings.